chore(@clayui/core): move story file out of source
Stories files were being compiled as part of the build process and we don't want to include them in the build, to avoid this we are moving out of the source and also removing the nested src from the components.
