+++ packages/clay-core/docs/provider.mdx
@@ -101,4 +101,4 @@ export const App = () => (
 
 ## API
 
-<div>[APITable "clay-core/src/provider/src/Provider.tsx"]</div>
+<div>[APITable "clay-core/src/provider/Provider.tsx"]</div>
