+++ packages/clay-core/src/provider/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-export * from './src/Provider';
+export * from './Provider';
